Updated CMS token and used shared test helper
