+++ HelpCentre/QA/tests/agent-greeting_test.exs
@@ -1,7 +1,7 @@
 defmodule AgentGreetingTest do
   use FlowTester.Case
   alias FlowTester.WebhookHandler, as: WH
-  alias FlowTester.WebhookHandler.Generic
+  alias HelpCentre.QA.Helpers
   # alias FlowTester.FlowStep, as: Step
   defp flow_path(flow_name), do: Path.join([__DIR__, "..", "flows_json", flow_name <> ".json"])
 
@@ -35,41 +35,41 @@ defmodule AgentGreetingTest do
   defp real_or_fake_cms(step, base_url, auth_token, :fake),
     do: WH.set_adapter(step, base_url, setup_fake_cms(auth_token))
 
-  defp turn_contacts_messages(env, ctx) do
-    assigned_to =
-      Map.get(ctx, :chat_assigned_to, %{
-        "id" => "some-uuid",
-        "name" => "Test Operator",
-        "type" => "OPERATOR"
-      })
+  # defp turn_contacts_messages(env, ctx) do
+  #   assigned_to =
+  #     Map.get(ctx, :chat_assigned_to, %{
+  #       "id" => "some-uuid",
+  #       "name" => "Test Operator",
+  #       "type" => "OPERATOR"
+  #     })
 
-    # IO.puts(inspect(assigned_to))
-    body = %{
-      "chat" => %{
-        "owner" => "+27821234567",
-        "state" => "OPEN",
-        "uuid" => "some-uuid",
-        "state_reason" => "Re-opened by inbound message.",
-        "assigned_to" => assigned_to,
-        "contact_uuid" => "some-uuid",
-        "permalink" => "https://whatsapp-praekelt-cloud.turn.io/app/c/some-uuid"
-      }
-    }
+  #   # IO.puts(inspect(assigned_to))
+  #   body = %{
+  #     "chat" => %{
+  #       "owner" => "+27821234567",
+  #       "state" => "OPEN",
+  #       "uuid" => "some-uuid",
+  #       "state_reason" => "Re-opened by inbound message.",
+  #       "assigned_to" => assigned_to,
+  #       "contact_uuid" => "some-uuid",
+  #       "permalink" => "https://whatsapp-praekelt-cloud.turn.io/app/c/some-uuid"
+  #     }
+  #   }
 
-    %Tesla.Env{env | status: 200, body: body}
-  end
+  #   %Tesla.Env{env | status: 200, body: body}
+  # end
 
-  defp setup_fake_turn(step, ctx) do
-    gen_pid = start_link_supervised!(Generic)
+  # defp setup_fake_turn(step, ctx) do
+  #   gen_pid = start_link_supervised!(Generic)
 
-    Generic.add_handler(
-      gen_pid,
-      ~r"/v1/contacts/[0-9]+/messages",
-      &turn_contacts_messages(&1, ctx)
-    )
+  #   Generic.add_handler(
+  #     gen_pid,
+  #     ~r"/v1/contacts/[0-9]+/messages",
+  #     &turn_contacts_messages(&1, ctx)
+  #   )
 
-    WH.set_adapter(step, "https://whatsapp-praekelt-cloud.turn.io/", Generic.wh_adapter(gen_pid))
-  end
+  #   WH.set_adapter(step, "https://whatsapp-praekelt-cloud.turn.io/", Generic.wh_adapter(gen_pid))
+  # end
 
   defp set_config(step) do
     step
@@ -82,8 +82,8 @@ defmodule AgentGreetingTest do
   end
 
   defp setup_flow(ctx) do
-    # When talking to real contentrepo, get the auth token from the API_TOKEN envvar.
-    auth_token = System.get_env("API_TOKEN", "CRauthTOKEN123")
+    # When talking to real contentrepo, get the auth token from the CMS_AUTH_TOKEN envvar.
+    auth_token = System.get_env("CMS_AUTH_TOKEN", "CRauthTOKEN123")
     kind = if auth_token == "CRauthTOKEN123", do: :fake, else: :real
 
     flow =
@@ -91,7 +91,7 @@ defmodule AgentGreetingTest do
       |> FlowTester.from_json!()
       |> real_or_fake_cms("https://content-repo-api-qa.prk-k8s.prd-p6t.org/", auth_token, kind)
       |> FlowTester.set_global_dict("settings", %{"contentrepo_qa_token" => auth_token})
-      |> setup_fake_turn(ctx)
+      |> Helpers.setup_fake_turn(ctx)
       |> set_config()
 
     %{flow: flow}
