+++ HelpCentre/QA/tests/intro-to-helpcentre_test.exs
@@ -3,6 +3,7 @@ defmodule IntroToHelpCentreTest do
   alias FlowTester.WebhookHandler, as: WH
   alias FlowTester.FlowStep
   alias FlowTester.WebhookHandler.Generic
+  alias HelpCentre.QA.Helpers
 
   defp flow_path(flow_name), do: Path.join([__DIR__, "..", "flows_json", flow_name <> ".json"])
 
@@ -215,57 +216,6 @@ defmodule IntroToHelpCentreTest do
   defp real_or_fake_cms(step, base_url, auth_token, :fake),
     do: WH.set_adapter(step, base_url, setup_fake_cms(auth_token))
 
-  defp turn_contacts_messages(env, ctx) do
-    assigned_to =
-      Map.get(ctx, :chat_assigned_to, %{
-        "id" => "some-uuid",
-        "name" => "Test Operator",
-        "type" => "OPERATOR"
-      })
-
-    # IO.puts(inspect(assigned_to))
-    body = %{
-      "messages" => [
-        %{
-          "id" => "someid",
-        }
-      ],
-      "chat" => %{
-        "owner" => "+27821234567",
-        "state" => "OPEN",
-        "uuid" => "some-uuid",
-        "state_reason" => "Re-opened by inbound message.",
-        "assigned_to" => assigned_to,
-        "contact_uuid" => "some-uuid",
-        "permalink" => "https://whatsapp-praekelt-cloud.turn.io/app/c/some-uuid"
-      }
-    }
-
-    # IO.puts(inspect(body))
-    %Tesla.Env{env | status: 200, body: body}
-  end
-
-  defp turn_add_label(env, _ctx) do
-    %Tesla.Env{env | status: 200}
-  end
-
-  defp setup_fake_turn(step, ctx) do
-    gen_pid = start_link_supervised!(Generic, id: :fake_turn)
-
-    Generic.add_handler(
-      gen_pid,
-      ~r"/v1/contacts/[0-9]+/messages",
-      &turn_contacts_messages(&1, ctx)
-    )
-
-    Generic.add_handler(
-      gen_pid,
-      "/v1/messages/someid/labels",
-      &turn_add_label(&1, ctx)
-    )
-
-    WH.set_adapter(step, "https://whatsapp-praekelt-cloud.turn.io/", Generic.wh_adapter(gen_pid))
-  end
 
   defp aaq_inbound_check(env, _ctx) do
     body = %{
@@ -343,8 +293,8 @@ defmodule IntroToHelpCentreTest do
   end
 
   defp setup_flow(ctx) do
-    # When talking to real contentrepo, get the auth token from the API_TOKEN envvar.
-    auth_token = System.get_env("API_TOKEN", "CRauthTOKEN123")
+    # When talking to real contentrepo, get the auth token from the CMS_AUTH_TOKEN envvar.
+    auth_token = System.get_env("CMS_AUTH_TOKEN", "CRauthTOKEN123")
     kind = if auth_token == "CRauthTOKEN123", do: :fake, else: :real
 
     flow =
@@ -352,7 +302,7 @@ defmodule IntroToHelpCentreTest do
       |> FlowTester.from_json!()
       |> real_or_fake_cms("https://content-repo-api-qa.prk-k8s.prd-p6t.org/", auth_token, kind)
       |> FlowTester.set_global_dict("settings", %{"contentrepo_qa_token" => auth_token})
-      |> setup_fake_turn(ctx)
+      |> Helpers.setup_fake_turn(ctx)
       |> setup_fake_aaq(ctx)
       |> set_config()
 
@@ -403,20 +353,7 @@ defmodule IntroToHelpCentreTest do
     |> FlowTester.start()
     |> receive_message(%{
       text: "*{MyHealth} Main Menu*\n\nTap the ‘Menu’ button to make your selection." <> _,
-      list:
-        {"Menu",
-         [
-           {"Your health guide 🔒", "Your health guide 🔒"},
-           {"View topics for you 📚", "View topics for you 📚"},
-           {"Chat to a nurse 🧑🏾‍⚕️", "Chat to a nurse 🧑🏾‍⚕️"},
-           {"Your profile ({0%}) 👤", "Your profile ({0%}) 👤"},
-           {"Manage updates 🔔", "Manage updates 🔔"},
-           {"Manage data 🖼️", "Manage data 🖼️"},
-           {"Help centre 📞", "Help centre 📞"},
-           {"Take a tour 🚌", "Take a tour 🚌"},
-           {"About and Privacy policy ℹ️", "About and Privacy policy ℹ️"},
-           {"Talk to a counsellor", "Talk to a counsellor"}
-         ]}
+      list: {"Menu", _}
     })
     |> FlowTester.send(button_label: "Help centre 📞")
     |> receive_message(%{
